requirements.txt
flask
gunicorn
pillow
qrcode[pil]
